Refactor _config.py, split remaining functions into ._diagnostics.py / ._utils.py
