+++ esmvalcore/_main.py
@@ -163,7 +163,7 @@ class Recipes():
         """
         import os
 
-        from ._config import DIAGNOSTICS_PATH
+        from ._diagnostics import DIAGNOSTICS_PATH
         from ._logging import configure_logging
         configure_logging(console_log_level='info')
         recipes_folder = os.path.join(DIAGNOSTICS_PATH, 'recipes')
@@ -193,7 +193,7 @@ class Recipes():
         import os
         import shutil
 
-        from ._config import DIAGNOSTICS_PATH
+        from ._diagnostics import DIAGNOSTICS_PATH
         from ._logging import configure_logging
         configure_logging(console_log_level='info')
         installed_recipe = os.path.join(DIAGNOSTICS_PATH, 'recipes', recipe)
@@ -218,7 +218,7 @@ class Recipes():
         """
         import os
 
-        from ._config import DIAGNOSTICS_PATH
+        from ._diagnostics import DIAGNOSTICS_PATH
         from ._logging import configure_logging
         configure_logging(console_log_level='info')
         installed_recipe = os.path.join(DIAGNOSTICS_PATH, 'recipes', recipe)
@@ -316,7 +316,7 @@ class ESMValTool():
         import os
         import shutil
 
-        from ._config import DIAGNOSTICS_PATH
+        from ._diagnostics import DIAGNOSTICS_PATH
         from ._logging import configure_logging
 
         if config_file:
