+++ esmvalcore/_utils.py
@@ -1,29 +1,15 @@
-"""ESMValTool configuration."""
+"""ESMValTool utilities."""
 import logging
 import os
-from pathlib import Path
 
 import yaml
 
+from ._diagnostics import DIAGNOSTICS_PATH
 from .cmor.table import CMOR_TABLES
 
 logger = logging.getLogger(__name__)
 
-
-def find_diagnostics():
-    """Try to find installed diagnostic scripts."""
-    try:
-        import esmvaltool
-    except ImportError:
-        return Path.cwd()
-    # avoid a crash when there is a directory called
-    # 'esmvaltool' that is not a Python package
-    if esmvaltool.__file__ is None:
-        return Path.cwd()
-    return Path(esmvaltool.__file__).absolute().parent
-
-
-DIAGNOSTICS_PATH = find_diagnostics()
+TAGS_CONFIG_FILE = DIAGNOSTICS_PATH / 'config-references.yml'
 
 
 def get_institutes(variable):
@@ -48,9 +34,6 @@ def get_activity(variable):
         return None
 
 
-TAGS_CONFIG_FILE = os.path.join(DIAGNOSTICS_PATH, 'config-references.yml')
-
-
 def _load_tags(filename=TAGS_CONFIG_FILE):
     """Load the reference tags used for provenance recording."""
     if os.path.exists(filename):
@@ -61,9 +44,6 @@ def _load_tags(filename=TAGS_CONFIG_FILE):
         # This happens if no diagnostics are installed
         logger.debug("No tags loaded, file %s not present", filename)
         return {}
-
-
-TAGS = _load_tags()
 
 
 def get_tag_value(section, tag):
@@ -81,3 +61,6 @@ def get_tag_value(section, tag):
 def replace_tags(section, tags):
     """Replace a list of tags with their values."""
     return tuple(get_tag_value(section, tag) for tag in tags)
+
+
+TAGS = _load_tags()
