+++ tests/integration/test_recipe.py
@@ -1214,7 +1214,7 @@ def test_diagnostic_task_provenance(
     monkeypatch,
     config_user,
 ):
-    monkeypatch.setattr(esmvalcore._config, 'TAGS', TAGS)
+    monkeypatch.setattr(esmvalcore._utils, 'TAGS', TAGS)
     monkeypatch.setattr(esmvalcore._recipe, 'TAGS', TAGS)
     monkeypatch.setattr(esmvalcore._task, 'TAGS', TAGS)
